ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic

requests

google-genai
langchain
langchain-google-genai

numpy
scikit-learn
textblob
nltk

# optional but useful for stability
tenacity